get the http types there own type file
