+++ src/bp/nlu/stan/api.ts
@@ -11,7 +11,8 @@ import modelIdService from 'nlu/engine/model-id-service'
 import { authMiddleware, handleErrorLogging, handleUnexpectedError } from '../../http-utils'
 import Logger from '../../simple-logger'
 
-import { PredictOutput, TrainInput, http } from '../typings_v1'
+import { PredictOutput, TrainInput } from '../typings_v1'
+import * as http from './http-typings'
 import { ModelRepoOptions, ModelRepository } from './model-repo'
 import TrainService from './train-service'
 import TrainSessionService from './train-session-service'
